+++ ExtensionMethods.cs
@@ -5,4 +5,5 @@ namespace Valheim.ServersideQoL;
 
 static class ExtensionMethods
 {
+    public static ExtendedZDO? GetExtendedZDO(this ZDOMan instance, ZDOID id) => (ExtendedZDO?)instance.GetZDO(id);
 }